==> 05_sim_nonlinear_continuous.R <==
conflicted::conflict_prefer("filter", "dplyr")
conflicted::conflict_prefer("unpack", "tidyr")
conflicted::conflicts_prefer(purrr::set_names)

here::i_am("Scripts R/05_sim_nonlinear_continuous.R")

# Libraries
library(tidyverse)
library(patchwork)
library(grid)
library(mvtnorm)
library(cubature)
library(nlme)
library(furrr)
library(here)

setwd(here())

n_cores <- min(parallel::detectCores() - 2, 30)
options(mc.cores = n_cores)
plan(multicore)

# Some ggplot options
theme_set(theme_bw())
theme_update(text = element_text(family = "Linux Biolinum O", size = 20),
             plot.title = element_text(face = "bold", hjust = 0.5))

## ------------------ Setting parameters

## Setting up a seed for the simulation
set.seed(2021)

## General parameters
# Number of simulations
n_sim   <- 1000
# Number of individuals
n_ind   <- 200
# Number of replicates
n_env   <- 10

## Sigmoidal reaction norm parameters
avg_sig <-
    c(a = 1,
      b = 3)
mat_sigm <- matrix(c(0.1, 0.02, 0.02, 0.05), ncol = 2)
sigmoid <- function(x, pars) {
    pars[ , "a"] / (1 + exp(-(pars[ , "b"] * x)))
}
sigmoid_cuba <- function(e, pars) {
    pars[1, ] / (1 + exp(-(pars[2, ] * e)))
}
d_theta_sigmoid_cuba <- function(e, pars) {
    matrix(c(
        1 / (1 + exp(-(pars[2, ] * e))),
        pars[1, ] * exp(-(pars[2, ] * e)) * e / (1 + exp(-(pars[2, ] * e)))^2
    ),
           nrow = 2,
           byrow = TRUE)
}
d_env_sigmoid_cuba <- function(e, pars) {
    pars[1, ] * (exp(-pars[2, ] * e) * pars[2, ])/(1 + exp(-pars[2, ] * e))^2
}
dd_env_sigmoid_cuba <- function(e, pars) {
    (pars[1, ] * pars[2, ]^2 * exp(-pars[2, ] * e) *
     (exp(-pars[2, ] * e)^2 - 1)) /
        (exp(-pars[2, ] * e)^4 + 4 * exp(-pars[2, ] * e)^3 +
         6 * exp(-pars[2, ] * e)^2 + 4 * exp(-pars[2, ] * e) + 1)
}

## Gaussian-Gompertz reaction norm parameter
avg_gg <-
    c(cmax       = 1,
      xopt       = 0.9,
      rho        = 8,
      sigma_gaus = 0.4)
mat_gausgompz <- matrix(c(0.1,      0.01,   0,          0,
                          0.01,     0.05,   0,          0,
                          0,        0,      0,          0,
                          0,        0,      0,          0),
                        ncol = 4)
gausgompz <- function(x, pars) {
    pars[ , "cmax"] * exp(
        - exp(pars[ , "rho"] * (x - pars[ , "xopt"]) - 6) -       # Gompertz part
            pars[ , "sigma_gaus"] * (x - pars[ , "xopt"])^2       # Gaussian part
    )
}
gausgompz_cuba <- function(e, pars) {
    pars[1, ] * exp(
        - exp(pars[3, ] * (e - pars[2, ]) - 6) -        # Gompertz part
            pars[4, ] * (e - pars[2, ])^2               # Gaussian part
    )
}
d_theta_gausgompz_cuba <- function(e, pars) {
    matrix(c(
            exp(-exp(-6 - pars[2, ] * pars[3, ] + e * pars[3, ]) - (pars[2, ] - e)^2 * pars[4, ]),
            pars[1, ] * exp(-exp(-6 - pars[2, ] * pars[3, ] + e * pars[3, ]) - (pars[2, ] - e)^2 * pars[4, ]) * (exp(-6 - pars[2, ] * pars[3, ] + e * pars[3, ]) * pars[3, ] + 2 * (-pars[2, ] + e) * pars[4, ])#,
#             pars[1, ] * exp(-6 - exp(-6 - pars[2, ] * pars[3, ] + e * pars[3, ]) + (-pars[2, ] + e) * pars[3, ] - (pars[2, ] - e)^2 * pars[4, ]) * (pars[2, ] - e),
#             -pars[1, ] * exp(-exp(-6 - pars[2, ] * pars[3, ] + e * pars[3, ]) - (pars[2, ] - e)^2 * pars[4, ]) * (pars[2, ] - e)^2
    ),
           nrow = 2,
           byrow = TRUE)
}
d_env_gausgompz_cuba <- function(e, pars) {
    -pars[1, ] * (pars[3, ] * exp(pars[3, ] * (e - pars[2, ]) - 6) +
                  2 * pars[4, ] * (e - pars[2, ])) *
        exp(-(exp(pars[3, ] * (e - pars[2, ]) - 6) + pars[4, ] * (e - pars[2, ])^2))
}
dd_env_gausgompz_cuba <- function(e, pars) {
    pars[1, ] * (pars[3, ] * exp(pars[3, ] * (e - pars[2, ]) - 6) +
                 2 * pars[4, ] * (e - pars[2, ]))^2 *
        exp(-(exp(pars[3, ] * (e - pars[2, ]) - 6) + pars[4, ] * (e - pars[2, ])^2)) -
        pars[1, ] * (exp(pars[3, ] * (e - pars[2, ]) - 6) * pars[3, ]^2 + 2 * pars[4, ]) *
        exp(-(exp(pars[3, ] * (e - pars[2, ]) - 6) + pars[4, ] * (e - pars[2, ])^2))
}
fixed_gg <- which(diag(mat_gausgompz) == 0)
mat_gausgompz_nofixed <- mat_gausgompz[-fixed_gg, -fixed_gg]

## Residuals
sd_res <- 0.01

## Environment moments from norm(0, 1)
mean_env <- 0
mean_env2 <- 1
mean_env3 <- 0
mean_env4 <- 3
var_env <- 1
var_env2 <- 2
M_env <- matrix(c(1, mean_env, mean_env2,
                  mean_env, mean_env2, mean_env3,
                  mean_env2, mean_env3, mean_env4),
                nrow = 3, ncol = 3)
env_VCV <- matrix(c(0, 0, 0, 0, var_env, 0, 0, 0, var_env2), nrow = 3, ncol = 3)

# Generating a "template" for environment following a "perfect" normal distribution
N_template <- 9999
env_template <-
    seq(1/(N_template+1),
        N_template / (N_template + 1),
        length.out = N_template) |>
    qnorm() |>
    scale() |>
    as.vector()
env_seq <- seq(-2.7, 2.7, length.out = 1000)

## ------------------ Functions for the simulations

## Function to simulate the reaction norm, with random variations
# Args: - n_sim: Number of simulations
#       - n_ind: Number of genotypes
#       - n_env: Number of individuals
#       - avg_pars: average parameters to be used for reaction norm (must be named)
#       - mat_pars: VCV matrix of the parameters
#       - shape_fn: function giving the shape of the reaction norm
# Value: A tbl with reaction norm for each env value (N * n_env)
simulate_rn <- function(n_sim, n_ind, n_env, avg_pars, mat_pars, shape_fn) {
    # Generating the parameters for each indivudal
    pars <- rmvnorm(n_sim * n_ind, avg_pars, mat_pars)
    pars <- pars[rep(1:nrow(pars), each = n_env), ]

    # Generating the environment for each measure
    env <- rnorm(n_sim * n_ind * n_env)

    # Generating the reaction norms, then the individual phenotypes and formatting
    out <-
        tibble(Simulation = as_factor(rep(1:n_sim, each = n_ind * n_env)),
               Individual = as_factor(rep(rep(1:n_ind, each = n_env), n_sim)),
               Env        = env,
               Phen       = shape_fn(env, pars) +
                            rnorm(n_sim * n_ind * n_env, 0, sd_res)) |>
        mutate(Env = as.vector(scale(Env)),
               .by = Simulation) |>
        nest_by(Simulation, .key = "Data") |>
        ungroup()

    return(out)
}

## Function to run the non-linear model on the reaction norm data using nlme
# Args: - df: A dataset containing the reaction norm data
#       - shape: the function of the true reaction norm
# Value: The model fit for the data
fit_nlme <- function(df, shape) {
    if (shape == "Sigmoid") {
        nlme(Phen ~ I(a / (1 + exp(-(b * Env)))),
             fixed  = a + b ~ 1,
             random = a + b ~ 1|Individual,
             data   = df,
             start  = avg_sig)
    } else {
        nlme(Phen ~ I(cmax * exp(- exp(rho * (Env - xopt) - 6) -
                                 sigma_gaus * (Env - xopt)^2)),
             fixed  = cmax + xopt + rho + sigma_gaus ~ 1,
             random = cmax + xopt ~ 1|Individual,
             data   = df,
             start  = avg_gg[c("cmax", "xopt", "rho", "sigma_gaus")])
    }
}

## Compute the G matrix from variances and correlation
# Args: - vars: vector of variances (diagonal elements of G)
#       - corr: correlation coefficient (only one as G are all dim 2)
# Value: the G-matrix
vc_to_vcv <- function(vars, corr) {
   # Removing the residual variance and generating a diagonal matrix from variances
   G <- diag(vars[-3])

   # Adding the covariance
   G[1, 2] <- G[2, 1] <- corr * sqrt(G[1, 1] * G[2, 2])

   return(G)
}

## Compute the average conditional to the environment (E_g_e)
# Args: - shape: the function of the reaction norm fitted by the model
#       - theta: the average parameters estimated by the model
#       - G_theta: the genetic variance-covariance matrix estimated by the model
#       - e: the environmental value to condition to
#       - width: the width over which the integral must be computed (10 is a generally a good value)
# Value: The value for E_g_e (numeric)
compute_Eg_e <- function(shape, theta, G_theta, e, width = 10, fixed = NA) {
    # This function is modified from the QGglmm package

    # Handling when some terms are fixed
    if (!any(is.na(fixed))) {
        var       <- setdiff(1:length(theta), fixed)
        full_theta <- theta
        var_theta  <- theta[-fixed]
        if (nrow(G_theta) == length(theta)) {
            G_theta <- G_theta[-fixed, -fixed]
        }
    } else {
        full_theta <- theta
        var_theta  <- theta
    }

    # Setting the integral width according to vcov (lower mean-w, upper mean+w)
    w <- sqrt(diag(G_theta)) * width

    # Number of dimensions
    d <- length(w)

    # Computing the logdet of vcov
    logdet <- calc_logdet(G_theta)

    # Average
    avg <- cubature::hcubature(
        f  = function(x) {
            full_x      <- matrix(full_theta, nrow = length(full_theta), ncol = ncol(x))
            if (!any(is.na(fixed))) { full_x[var, ] <- x } else { full_x <- x }
            shape(e, full_x) * vec_mvnorm(x, var_theta, G_theta, logdet)
        },
        lowerLimit = var_theta - w,
        upperLimit = var_theta + w,
        fDim       = 1,
        tol        = 0.001,
        absError   = 0.0001,
        vectorInterface = TRUE
    )$integral

    return(avg)
}

## Compute the plastic variance (V_plas)
# Args: - shape: the name of the shape used to simulate/estimate the RN
#       - theta: the average parameters estimated by the model
#       - vars: the vars estimated from the model
#       - env: the environmental values over which the model has been estimated
#       - width: the width over which the integral must be computed (10 is a generally a good value)
#       - fixed: which part of the parameters are fixed (no genetic variation)
#       - average: should the average of variances be returned?
#                  If FALSE, return the variance for each environmental value
# Value: The value for V_plas (numeric)
compute_vplas <- function(shape, theta, G_theta, env, width = 10, fixed = NA) {
    # Defining the shape function
    if (shape == "Sigmoid") {
        shape_func <- sigmoid_cuba
    } else {
        shape_func <- gausgompz_cuba
    }

    map_dbl(env,
            \(e) compute_Eg_e(e = e, shape = shape_func, theta = theta, G_theta = G_theta,
                              width = width, fixed = fixed),
            .progress = TRUE) |>
        rep(n_ind * n_env) |>
        var()
}

## Compute the slope variance (V_Sl component of V_plas)
# Args: - shape: the name of the shape used to simulate/estimate the RN
#       - theta: the average parameters estimated by the model
#       - vars: the vars estimated from the model
#       - env: the environmental values over which the model has been estimated
#       - width: the width over which the integral must be computed (10 is a generally a good value)
#       - fixed: which part of the parameters are fixed (no genetic variation)
#       - average: should the average of variances be returned?
#                  If FALSE, return the variance for each environmental value
# Value: The value for V_sl (numeric)
compute_v_sl <- function(shape, theta, G_theta, env, width = 10, fixed = NA) {
    # Defining the shape function
    if (shape == "Sigmoid") {
        shape_func <- d_env_sigmoid_cuba
    } else {
        shape_func <- d_env_gausgompz_cuba
    }

    mean_sl <-
        map_dbl(env,
            \(e) compute_Eg_e(e = e, shape = shape_func, theta = theta, G_theta = G_theta,
                              width = width, fixed = fixed),
            .progress = TRUE) |>
        mean()

    var_sl <- (mean_sl)^2 * var(env)

    return(var_sl)
}

## Compute the curvature variance (V_Cv component of V_plas)
# Args: - shape: the name of the shape used to simulate/estimate the RN
#       - theta: the average parameters estimated by the model
#       - vars: the vars estimated from the model
#       - env: the environmental values over which the model has been estimated
#       - width: the width over which the integral must be computed (10 is a generally a good value)
#       - fixed: which part of the parameters are fixed (no genetic variation)
#       - average: should the average of variances be returned?
#                  If FALSE, return the variance for each environmental value
# Value: The value for V_cv (numeric)
compute_v_cv <- function(shape, theta, G_theta, env, width = 10, fixed = NA) {
    # Defining the shape function
    if (shape == "Sigmoid") {
        shape_func <- dd_env_sigmoid_cuba
    } else {
        shape_func <- dd_env_gausgompz_cuba
    }

    mean_cv <-
        map_dbl(env,
                \(e) compute_Eg_e(e = e, shape = shape_func, theta = theta, G_theta = G_theta,
                                  width = width, fixed = fixed),
                .progress = TRUE) |>
        mean()

    var_cv <- 0.25 * (mean_cv)^2 * var(env^2)

    return(var_cv)
}

## Log-determinant of a VCV matrix
calc_logdet <- function(Sigma) {
    sum(log(eigen(Sigma, symmetric = TRUE, only.values = TRUE)$values))
}

## Vectorised multivariate Gaussian density function
# Shamelessly stolen from cubature vignette (credit to Balasubramanian Narasimhan)
# logdet = sum(log(eigen(sd_res, symmetric = TRUE, only.values = TRUE)$values))
vec_mvnorm <- function(x, mean, Sigma, logdet = NULL) {
    # If logdet not provided, compute it
    if (is.null(logdet)) {
        logdet <- calc_logdet(Sigma)
    }
    # Compute Mahalanobis distance (corresponds to the exp. part of the density)
    distval <- stats::mahalanobis(t(x), center = mean, cov = Sigma)
    # Compute the vectorised MVN density
    out <- exp(matrix(-(nrow(x) * log(2 * pi) + logdet + distval)/2, ncol = ncol(x)))
    return(out)
}


## Compute the genetic variance conditionnally to the environment (V_g_e)
# Args: - shape: the function of the reaction norm fitted by the model
#       - theta: the average parameters estimated by the model
#       - G_theta: the genetic variance-covariance matrix estimated by the model
#       - e: the environmental value to condition to
#       - width: the width over which the integral must be computed (10 is a generally a good value)
# Value: The value for V_g_e (numeric)
compute_vg_e <- function(shape, theta, G_theta, e, width = 10, fixed = NA) {
    # This function is modified from the QGglmm package

    # Handling when some terms are fixed
    if (!any(is.na(fixed))) {
        var       <- setdiff(1:length(theta), fixed)
        full_theta <- theta
        var_theta  <- theta[-fixed]
        if (nrow(G_theta) == length(theta)) {
            G_theta <- G_theta[-fixed, -fixed]
        }
    } else {
        full_theta <- theta
        var_theta  <- theta
    }

    # Setting the integral width according to vcov (lower mean-w, upper mean+w)
    w <- sqrt(diag(G_theta)) * width

    # Number of dimensions
    d <- length(w)

    # Computing the logdet of vcov
    logdet <- calc_logdet(G_theta)

    # Average
    avg <- cubature::hcubature(
        f  = function(x) {
            full_x      <- matrix(full_theta, nrow = length(full_theta), ncol = ncol(x))
            if (!any(is.na(fixed))) { full_x[var, ] <- x } else { full_x <- x }
            shape(e, full_x) * vec_mvnorm(x, var_theta, G_theta, logdet)
        },
        lowerLimit = var_theta - w,
        upperLimit = var_theta + w,
        fDim       = 1,
        tol        = 0.001,
        absError   = 0.0001,
        vectorInterface = TRUE
    )$integral

    # Computing the integral
    cubature::hcubature(
        f  = function(x) {
            full_x      <- matrix(full_theta, nrow = length(full_theta), ncol = ncol(x))
            if (!any(is.na(fixed))) { full_x[var, ] <- x } else { full_x <- x }
            (shape(e, full_x) - avg)^2 * vec_mvnorm(x, var_theta, G_theta, logdet)
        },
        lowerLimit = var_theta - w,
        upperLimit = var_theta + w,
        fDim       = 1,
        tol        = 0.001,
        absError   = 0.0001,
        vectorInterface = TRUE
    )$integral
}

## Compute the genetic variance (V_gen)
# Args: - shape: the name of the shape used to simulate/estimate the RN
#       - theta: the average parameters estimated by the model
#       - vars: the vars estimated from the model
#       - env: the environmental values over which the model has been estimated
#       - width: the width over which the integral must be computed (10 is a generally a good value)
#       - fixed: which part of the parameters are fixed (no genetic variation)
#       - average: should the average of variances be returned?
#                  If FALSE, return the variance for each environmental value
# Value: The value for V_gen (numeric)
compute_vgen <- function(shape, theta, G_theta, env, width = 10, fixed = NA, average = TRUE) {
    # Defining the shape function
    if (shape == "Sigmoid") {
        shape_func <- sigmoid_cuba
    } else {
        shape_func <- gausgompz_cuba
    }

    out <-
        map_dbl(env,
        \(e) compute_vg_e(e = e, shape = shape_func, theta = theta, G_theta = G_theta,
                            width = width, fixed = fixed),
        .progress = TRUE)

    if (average) { out <- mean(out) }

    return(out)
}

## Compute the gradient psi conditionnally to the environment (psi_e)
# Args: - shape: the function of the reaction norm fitted by the model
#       - theta: the average parameters estimated by the model
#       - G_theta: the genetic variance-covariance matrix estimated by the model
#       - e: the environmental value to condition to
#       - width: the width over which the integral must be computed (10 is a generally a good value)
# Value: The value for psi_e (numeric)
compute_psi_e <- function(shape, theta, G_theta, e, width = 10, fixed = NA) {
    # This function is modified from the QGglmm package

    # Handling when some terms are fixed
    if (!any(is.na(fixed))) {
        var       <- setdiff(1:length(theta), fixed)
        full_theta <- theta
        var_theta  <- theta[-fixed]
        if (nrow(G_theta) == length(theta)) {
            G_theta <- G_theta[-fixed, -fixed]
        }
    } else {
        full_theta <- theta
        var_theta  <- theta
    }

    # Setting the integral width according to vcov (lower mean-w, upper mean+w)
    w <- sqrt(diag(G_theta)) * width

    # Number of dimensions
    d <- length(w)

    # Computing the logdet of vcov
    logdet <- calc_logdet(G_theta)

    # Average
    k <- shape(e, matrix(full_theta, nrow = length(full_theta)))

    # Computing the integral for psi
    psi <- cubature::hcubature(
        f  = function(x) {
            full_x      <- matrix(full_theta, nrow = length(full_theta), ncol = ncol(x))
            if (!any(is.na(fixed))) { full_x[var, ] <- x } else { full_x <- x }
            shape(e, full_x) * matrix(rep(vec_mvnorm(x, var_theta, G_theta, logdet), d),
                                      nrow = d,
                                      byrow = TRUE)
        },
        lowerLimit = var_theta - w,
        upperLimit = var_theta + w,
        fDim       = d,
        tol        = 0.001,
        absError   = 0.0001,
        vectorInterface = TRUE
    )$integral

    return(psi)
}

## Compute the gradient (psi)
# Args: - shape: the name of the shape used to simulate/estimate the RN
#       - theta: the average parameters estimated by the model
#       - vars: the vars estimated from the model
#       - env: the environmental values over which the model has been estimated
#       - width: the width over which the integral must be computed (10 is a generally a good value)
#       - fixed: which part of the parameters are fixed (no genetic variation)
#       - average: should the average of variances be returned?
#                  If FALSE, return the variance for each environmental value
# Value: A matrix of the values of psi for each parameter (col) and environment (row) (numeric)
compute_psi <- function(shape, theta, G_theta, env, width = 10, fixed = NA, average = TRUE) {
    # Defining the shape function
    if (shape == "Sigmoid") {
        shape_func <- d_theta_sigmoid_cuba
    } else {
        shape_func <- d_theta_gausgompz_cuba
    }

    if (all(is.na(fixed))) {
        names <- names(theta)
    } else {
        names <- names(theta[-fixed])
    }

    out <-
        map(env,
            \(e) {
                compute_psi_e(e = e, shape = shape_func, theta = theta, G_theta = G_theta,
                              width = width, fixed = fixed) |>
                    set_names(names) |>
                    as.list() |>
                    as_tibble()
            },
            .progress = TRUE) |>
        list_rbind() |>
        as.matrix()

    return(out)
}

## Compute the total additive genetic variance (V_add)
# Args: - psi: a matrix containing the gradient for all parameters and environments
#       - mat_g: the G matrix
# Value: The value for V_Add and the gamma coefficients (numeric)
compute_vadd <- function(psi, mat_g, average = TRUE) {
    out <-
        tibble(V_A      = diag(psi %*% mat_g %*% t(psi)),
               Gamma_1  = as.numeric(psi[ , 1]^2 * mat_g[1, 1]),
               Gamma_2  = as.numeric(psi[ , 2]^2 * mat_g[2, 2]),
               Gamma_12 = as.numeric(2 * psi[ , 1] * psi[ , 2] * mat_g[1, 2]))

    if (average) {

        out <-
            out |>
            summarise(V_Add = mean(V_A),
                      across(starts_with("Gamma"), mean)) |>
            mutate(across(starts_with("Gamma"), \(v) { v / V_Add }))

        return(out)

    } else {
        out <-
            out |>
            mutate(across(starts_with("Gamma"), \(v) { v / V_A }))

        return(out)
    }
}

## Compute the total additive genetic variance (V_A)
# Args: - psi: a matrix containing the gradient for all parameters and environments
#       - mat_g: the G matrix
# Value: The value for V_A (numeric)
compute_va <- function(psi, mat_g) {
    out <- (t(colMeans(psi)) %*% mat_g %*% colMeans(psi)) |>
           as.vector()

    return(out)
}

## Compute the total additive genetic variance (V_add)
# Args: - psi: a matrix containing the gradient for all parameters and environments
#       - mat_g: the G matrix
# Value: The value for V_Add and the gamma coefficients (numeric)
compute_vaxe <- function(psi, mat_g) {
    cov_psi <- cov(psi)

    out <-
        tibble(V_AxE   = sum(cov_psi * mat_g),
               Iota_1  = cov_psi[1, 1] * mat_g[1, 1],
               Iota_2  = cov_psi[2, 2] * mat_g[2, 2],
               Iota_12 = 2 * cov_psi[1, 2] * mat_g[1, 2]) |>
        summarise(V_AxE = mean(V_AxE),
                  across(starts_with("Iota"), mean)) |>
        mutate(across(starts_with("Iota"), \(v) { v / V_AxE }))

    return(out)
}

## Set colours for graphics depending on the type of variance
# Args: - vec_pars: a vector of parameter names
# Value: A vector of HTML colour names corresponding to the names in the input (character)
set_colours <- function(vec_pars) {
    case_when(
        vec_pars == "V[Plas]"      ~ "#AA0000",
        vec_pars == "P[RN]^2"      ~ "#AA0000",
        str_detect(vec_pars, "π")  ~ "#AA0000",
        vec_pars == "V[Gen]"       ~ "#AA5500",
        vec_pars == "V[G,ε]"       ~ "#AA5500",
        vec_pars == "H[RN]^2"      ~ "#AA5500",
        vec_pars == "V[Add]"       ~ "#68349C",
        vec_pars == "V[A,ε]"       ~ "#68349C",
        vec_pars == "h[RN]^2"      ~ "#68349C",
        str_detect(vec_pars, "γ")  ~ "#68349C",
        vec_pars == "V[A]"         ~ "#228B22",
        vec_pars == "h^2"          ~ "#228B22",
        vec_pars == "V[AxE]"       ~ "#2D489A",
        vec_pars == "h[I]^2"       ~ "#2D489A",
        str_detect(vec_pars, "ι")  ~ "#2D489A"
    )
}

## ------------------ Simulating the data

# Generating the datasets
tbl_sim <-
    bind_rows(
        bind_cols(Shape = "Sigmoid",
                  simulate_rn(n_sim    = n_sim,
                              n_ind    = n_ind,
                              n_env    = n_env,
                              avg_pars = avg_sig,
                              mat_pars = mat_sigm,
                              shape_fn = sigmoid)),
        bind_cols(Shape = "GausGompz",
                  simulate_rn(n_sim    = n_sim,
                              n_ind    = n_ind,
                              n_env    = n_env,
                              avg_pars = avg_gg,
                              mat_pars = mat_gausgompz,
                              shape_fn = gausgompz)),
    )

## ------------------ Running the statistical models

## Running the models
tbl_sim[["Mod"]] <- future_map2(tbl_sim[["Data"]],
                                tbl_sim[["Shape"]],
                                \(df, sh) { fit_nlme(df, sh) },
                                .progress = TRUE)

## Getting the estimates
tbl_sim <-
    tbl_sim |>
    mutate(Param = list(fixed.effects(Mod[[1]])),
           Vars  = VarCorr(Mod[[1]])[ , "Variance"] |>
                   as.numeric() |>
                   set_names(case_when(
                       Shape[[1]] == "Sigmoid" ~ c("V_a", "V_b", "V_R"),
                       Shape[[1]] == "GausGompz" ~ c("V_cmax", "V_xopt", "V_R"),
                   )) |>
                   list(),
           Corr  = VarCorr(Mod[[1]])[2 , "Corr"] |>
                   as.numeric(),
           .by = c(Shape, Simulation))

## Saving parameters into RN functions to plot the curves later
# Sigmoid
fe_sig <- fixef(tbl_sim[["Mod"]][[1]])
func_sig_1 <- function(x) {
    fe_sig["a"] / (1 + exp(- fe_sig["b"] * x))
}
# Gaussian-Gompertz
fe_gg <- fixef(tbl_sim[["Mod"]][[n_sim + 1]])
func_gg_1 <- function(x) {
    fe_gg["cmax"] * exp(
        - exp(fe_gg["rho"] * (x - fe_gg["xopt"]) - 6) -       # Gompertz part
            fe_gg["sigma_gaus"] * (x - fe_gg["xopt"])^2       # Gaussian part
    )
}

## Saving the results (without the models)
tbl_sim <-
    tbl_sim |>
    select(-Mod)
saveRDS(tbl_sim, file = here("sims_nl_cont.rds"))
saveRDS(fe_sig, file = here("fe_sig_nl_cont.rds"))
saveRDS(func_sig_1, file = here("fun_sig_nl_cont.rds"))
saveRDS(fe_gg, file = here("fe_gg_nl_cont.rds"))
saveRDS(func_gg_1, file = here("fun_gg_nl_cont.rds"))

## ------------------ Computing the variance components

# Computing the variance components
tbl_vars <-
    tbl_sim |>
    mutate(G        = future_map2(Vars, Corr, vc_to_vcv),
           Fixed    = if_else(Shape == "GausGompz", list(fixed_gg), list(NA)),
           V_plas   = future_pmap_dbl(list(shape     = Shape,
                                           theta     = Param,
                                           G_theta   = G,
                                           fixed     = Fixed,
                                           env       = map(Data, "Env")),
                                      compute_vplas,
                                      .options = furrr_options(seed = TRUE),
                                      .progress = TRUE),
           V_Sl     = future_pmap_dbl(list(shape     = Shape,
                                           theta     = Param,
                                           G_theta   = G,
                                           fixed     = Fixed,
                                           env       = map(Data, "Env")),
                                      compute_v_sl,
                                      .options = furrr_options(seed = TRUE),
                                      .progress = TRUE),
           Pi_Sl    = V_Sl / V_plas,
           V_Cv     = future_pmap_dbl(list(shape     = Shape,
                                           theta      = Param,
                                           G_theta    = G,
                                           fixed     = Fixed,
                                           env       = map(Data, "Env")),
                                      compute_v_cv,
                                      .options = furrr_options(seed = TRUE),
                                      .progress = TRUE),
           Pi_Cv    = V_Cv / V_plas,
           V_gen    = future_pmap_dbl(list(shape     = Shape,
                                       theta     = Param,
                                       G_theta   = G,
                                       fixed     = Fixed,
                                       env       = map(Data, "Env")),
                                  compute_vgen,
                                  .options = furrr_options(seed = TRUE),
                                  .progress = TRUE,
                                  average = TRUE),
           Psi      = future_pmap(list(shape     = Shape,
                                       theta     = Param,
                                       G_theta   = G,
                                       fixed     = Fixed,
                                       env       = map(Data, "Env")),
                                  compute_psi,
                                  .options = furrr_options(seed = TRUE),
                                  .progress = TRUE),
           V_Add    = future_map2(Psi, G, compute_vadd),
           V_A      = future_map2_dbl(Psi, G, compute_va),
           V_AxE    = future_map2(Psi, G, compute_vaxe),
           V_res    = map_dbl(Vars, 3),
           V_tot    = V_plas + V_gen + V_res,
           V_phen   = map_dbl(Data, \(df) { var(df[["Phen"]]) })) |>
    select(Shape, Simulation, G, starts_with("V_")) |>
    unnest(c(V_Add, V_AxE)) |>
    mutate(P2       = V_plas / V_tot,
           H2_T     = V_gen / V_tot,
           h2_T     = V_Add / V_tot,
           h2_M     = V_A / V_tot,
           h2_I     = V_AxE / V_tot)

# Computing the true expected values
tbl_true_vars <-
    tibble(Shape    = c("Sigmoid", "GausGompz"),
           V_plas   = c(compute_vplas("Sigmoid", avg_sig, mat_sigm, env_template),
                        compute_vplas("GausGompz", avg_gg, mat_gausgompz, env_template,
                                      fixed = fixed_gg)),
           V_Sl     = c(compute_v_sl("Sigmoid", avg_sig, mat_sigm, env_template),
                        compute_v_sl("GausGompz", avg_gg, mat_gausgompz, env_template,
                                      fixed = fixed_gg)),
           Pi_Sl    = V_Sl / V_plas,
           V_Cv     = c(compute_v_cv("Sigmoid", avg_sig, mat_sigm, env_template),
                        compute_v_cv("GausGompz", avg_gg, mat_gausgompz, env_template,
                                     fixed = fixed_gg)),
           Pi_Cv    = V_Cv / V_plas,
           V_gen    = c(compute_vgen("Sigmoid", avg_sig, mat_sigm, env_template,
                                        average = TRUE),
                           compute_vgen("GausGompz", avg_gg, mat_gausgompz, env_template,
                                        fixed = fixed_gg, average = TRUE)),
           Psi      = list(compute_psi("Sigmoid", avg_sig, mat_sigm, env_template,
                                        average = TRUE),
                           compute_psi("GausGompz", avg_gg, mat_gausgompz, env_template,
                                        fixed = fixed_gg, average = TRUE)),
           V_Add    = list(compute_vadd(Psi[[1]], mat_sigm),
                           compute_vadd(Psi[[2]], mat_gausgompz_nofixed)),
           V_A      = c(compute_va(Psi[[1]], mat_sigm),
                           compute_va(Psi[[2]], mat_gausgompz_nofixed)),
           V_AxE    = list(compute_vaxe(Psi[[1]], mat_sigm),
                           compute_vaxe(Psi[[2]], mat_gausgompz_nofixed)),
           V_res    = sd_res^2,
           V_tot    = V_plas + V_gen + V_res) |>
    select(-Psi) |>
    unnest(c(V_Add, V_AxE)) |>
    mutate(P2       = V_plas / V_tot,
           H2_T     = V_gen / V_tot,
           h2_T     = V_Add / V_tot,
           h2_M     = V_A / V_tot,
           h2_I     = V_AxE / V_tot)

## Genetic variances across environments
# Getting variances across environments
tbl_vars_gamma_e <-
    tbl_sim |>
    mutate(G        = future_map2(Vars, Corr, vc_to_vcv),
           Fixed    = if_else(Shape == "GausGompz", list(fixed_gg), list(NA)),
           V_g_e    = future_pmap(list(shape     = Shape,
                                       theta     = Param,
                                       G_theta   = G,
                                       fixed     = Fixed),
                                  compute_vgen,
                                  .progress = TRUE,
                                  .options = furrr_options(seed = TRUE),
                                  average = FALSE,
                                  width = 9.5,
                                  env = env_seq),
           Psi      = future_pmap(list(shape     = Shape,
                                       theta     = Param,
                                       G_theta   = G,
                                       fixed     = Fixed),
                                  compute_psi,
                                  .options = furrr_options(seed = TRUE),
                                  .progress = TRUE,
                                  width = 9.5,
                                  env = env_seq),
           V_A_e    = future_map2(Psi, G, compute_vadd, average = FALSE)) |>
    mutate(Env = map(1:n(), \(x) { env_seq })) |>
    select(Shape, Env, V_g_e, V_A_e) |>
    unnest(c(Env, V_g_e, V_A_e))

tbl_vars_e <-
    tbl_vars_gamma_e |>
    select(-starts_with("Gamma")) |>
    rename(V_A_e = V_A) |>
    pivot_longer(starts_with("V"),
                 names_to = "Variance",
                 values_to = "Value") |>
    mutate(Variance = factor(Variance, levels = c("V_g_e", "V_A_e"))) |>
    summarise(Mean = mean(Value),
              Low  = quantile(Value, probs = 0.025),
              Up   = quantile(Value, probs = 0.975),
              .by  = c(Shape, Variance, Env))

tbl_gamma_e <-
    tbl_vars_gamma_e |>
    select(-starts_with("V")) |>
    pivot_longer(starts_with("Gamma"),
                 names_to = "Estimate",
                 values_to = "Value") |>
    mutate(Estimate = factor(Estimate, levels = c("Gamma_1", "Gamma_2", "Gamma_12"))) |>
    summarise(Mean = mean(Value),
              Low  = quantile(Value, probs = 0.025),
              Up   = quantile(Value, probs = 0.975),
              .by  = c(Shape, Estimate, Env))

# Correlation between V_tot and V_phen
tbl_vars |>
    summarise(Corr = cor(V_tot, V_phen),
              .by = Shape) |>
    deframe()

saveRDS(tbl_vars, file = here("vars_nl_cont.rds"))
saveRDS(tbl_true_vars, file = here("true_vars_nl_cont.rds"))
saveRDS(tbl_vars_e, file = here("vars_e_nl_cont.rds"))
saveRDS(tbl_gamma_e, file = here("gamma_e_nl_cont.rds"))

## ------------------ Graphical outputs

tbl_sim <- readRDS(here("sims_nl_cont.rds"))
tbl_vars <- readRDS(here("vars_nl_cont.rds"))
tbl_vars_e <- readRDS(here("vars_e_nl_cont.rds"))
tbl_gamma_e <- readRDS(here("gamma_e_nl_cont.rds"))
tbl_true_vars <- readRDS(here("true_vars_nl_cont.rds"))
fe_sig <- readRDS(file = here("fe_sig_nl_cont.rds"))
func_sig_1 <- readRDS(file = here("fun_sig_nl_cont.rds"))
fe_gg <- readRDS(file = here("fe_gg_nl_cont.rds"))
func_gg_1 <- readRDS(file = here("fun_gg_nl_cont.rds"))

options(scipen = 999)

# A data.frame with only the estimates
tbl_estim <-
    tbl_sim |>
    reframe(Coef = enframe(c(Param[[1]], Vars[[1]]),
                           name = "Param",
                           value = "Value"),
            .by = c(Shape, Simulation)) |>
    unpack(Coef) |>
    mutate(Value = as.numeric(Value))

## Raw parameter estimations
# Checking bias/precision on raw parameters

# Getting the relative errors on parameters
tbl_bias <-
    tbl_estim |>
    filter(Param != "V_R") |>
    mutate(
        Rel_Bias = case_when(
            Param == "a"            ~ (Value - avg_sig["a"]),
            Param == "b"            ~ (Value - avg_sig["b"]) / avg_sig["b"],
            Param == "V_a"          ~ (Value - mat_sigm[1, 1]) / mat_sigm[1, 1],
            Param == "V_b"          ~ (Value - mat_sigm[2, 2]) / mat_sigm[2, 2],
            Param == "cmax"         ~ (Value - avg_gg["cmax"]) / avg_gg["cmax"],
            Param == "xopt"         ~ (Value - avg_gg["xopt"]) / avg_gg["xopt"],
            Param == "rho"          ~ (Value - avg_gg["rho"]) / avg_gg["rho"],
            Param == "sigma_gaus"   ~ (Value - avg_gg["sigma_gaus"]) / avg_gg["sigma_gaus"],
            Param == "V_cmax"       ~ (Value - mat_gausgompz[1, 1]) / mat_gausgompz[1, 1],
            Param == "V_xopt"       ~ (Value - mat_gausgompz[2, 2]) / mat_gausgompz[2, 2],
            Param == "V_sig"        ~ (Value - mat_gausgompz[4, 4]) / mat_gausgompz[4, 4],
        ),
        Param =
            factor(Param,
                   levels = c("a", "b", "xopt", "cmax", "rho", "sigma_gaus",
                              "V_a", "V_b", "V_cmax", "V_xopt", "V_sig")) |>
                fct_recode("V[a]" = "V_a",
                           "V[b]" = "V_b",
                           "C[max]" = "cmax",
                           "x[opt]" = "xopt",
                           "sigma"  = "sigma_gaus",
                           "V[C[max]]" = "V_cmax",
                           "V[x[opt]]" = "V_xopt",
                           "V[sigma]"    = "V_sig"),
        Shape = recode(Shape, Sigmoid = "Sigmoid", GausGompz = "Thermal Perf. Curve") |>
                as_factor()
    )

# Plot of the relative errors
p_params <-
    ggplot(tbl_bias) +
    geom_violin(aes(x = Param, y = Rel_Bias), fill = "grey", scale = "width") +
    geom_hline(yintercept = 0) +
    stat_summary(aes(x = Param, y = Rel_Bias),
                 geom = "point",
                 colour = "red",
                 size = 2,
                 fun = "mean") +
    facet_grid(~ Shape, scale = "free_x", space = "free_x") +
    scale_x_discrete(labels = scales::parse_format()) +
    labs(x = "Variance component", y = "Relative Error") +
    theme(text         = element_text(family = "Linux Biolinum O"),
          axis.text.y  = element_text(size = 14),
          axis.text.x  = element_text(size = 14),
          axis.title   = element_text(size = 16),
          strip.text   = element_text(size = 20))

# Saving the plot
cairo_pdf(here("Figs/NonLin_Parameters_cont.pdf"), height = 5, width = 8)
plot(p_params)
dev.off()

# Checking for bias
tbl_bias |>
    summarise(P_val = wilcox.test(Rel_Bias)[["p.value"]],
              .by = c(Shape, Param)) |>
    pivot_wider(values_from = "P_val", names_from = "Param") |>
    mutate(across(-Shape, \(col) { num(col, digits = 5) })) |>
    print(width = Inf)

## Variance components
# Formatting the data
tbl_var_plot <-
    tbl_vars |>
    select(-G, -ends_with("_e"), -starts_with("V_")) |>
    pivot_longer(!c(Shape, Simulation),
                 names_to   = "Variance",
                 values_to  = "Value") |>
    left_join(tbl_true_vars |>
              select(-V_Sl, -V_Cv) |>
              pivot_longer(!Shape,
                           names_to  = "Variance",
                           values_to = "Truth")) |>
    mutate(Bias     = (Value - Truth),
           Rel_Bias = Bias / Truth,
           Variance = case_when(
               Variance == "V_plas"                          ~ "V[Plas]",
               Variance == "P2"                              ~ "P[RN]^2",
               Variance == "Pi_Sl"                           ~ "π[Sl]",
               Variance == "Pi_Cv"                           ~ "π[Cv]",
               Variance == "V_gen"                           ~ "V[Gen]",
               Variance == "H2_T"                            ~ "H[RN]^2",
               Variance == "V_A"                             ~ "V[A]",
               Variance == "h2_T"                            ~ "h[RN]^2",
               Variance == "h2_M"                            ~ "h^2",
               Variance == "h2_I"                            ~ "h[I]^2",
               Variance == "V_res"                           ~ "V[Res]",
               Variance == "V_tot"                           ~ "V[Tot]",
               Shape == "Sigmoid"   & Variance == "Gamma_1"  ~ "γ[L]",
               Shape == "Sigmoid"   & Variance == "Gamma_2"  ~ "γ[r]",
               Shape == "Sigmoid"   & Variance == "Gamma_12" ~ "γ[Lr]",
               Shape == "GausGompz" & Variance == "Gamma_1"  ~ "γ[C]",
               Shape == "GausGompz" & Variance == "Gamma_2"  ~ "γ[ε[0]]",
               Shape == "GausGompz" & Variance == "Gamma_12" ~ "γ[Cε[0]]",
               Shape == "Sigmoid"   & Variance == "Iota_1"   ~ "ι[L]",
               Shape == "Sigmoid"   & Variance == "Iota_2"   ~ "ι[r]",
               Shape == "Sigmoid"   & Variance == "Iota_12"  ~ "ι[Lr]",
               Shape == "GausGompz" & Variance == "Iota_1"   ~ "ι[C]",
               Shape == "GausGompz" & Variance == "Iota_2"   ~ "ι[ε[0]]",
               Shape == "GausGompz" & Variance == "Iota_12"  ~ "ι[Cε[0]]",
           ) |>
                      factor(levels = c("V[Plas]",
                                        "P[RN]^2",
                                        "π[Sl]",
                                        "π[Cv]",
                                        "V[Gen]",
                                        "H[RN]^2",
                                        "V[A]",
                                        "h[RN]^2",
                                        "γ[L]",
                                        "γ[r]",
                                        "γ[Lr]",
                                        "γ[C]",
                                        "γ[ε[0]]",
                                        "γ[Cε[0]]",
                                        "h^2",
                                        "h[I]^2",
                                        "ι[L]",
                                        "ι[r]",
                                        "ι[Lr]",
                                        "ι[C]",
                                        "ι[ε[0]]",
                                        "ι[Cε[0]]",
                                        "V[Res]",
                                        "V[Tot]")),
           Shape    = recode(Shape,
                             GausGompz = "Thermal Perf. Curve") |>
                      as_factor(),
           Colour = set_colours(Variance),
           Est_Type = if_else(str_detect(Variance, "^V"), "Variances", "Components") |>
                      factor(levels = c("Variances", "Components")))

# Now making the plot
p_var <-
    ggplot(tbl_var_plot) +
    geom_violin(aes(x = Variance, y = Bias), fill = "grey", scale = "width") +
    geom_hline(yintercept = 0) +
    stat_summary(aes(x = Variance, y = Bias),
                 geom = "point",
                 colour = "red",
                 size = 2,
                 fun = "mean") +
    facet_grid(~ Shape, scale = "free_x", space = "free_x") +
    scale_x_discrete(labels = scales::parse_format()) +
    labs(x = "Variance component", y = "Error") +
    theme(text         = element_text(family = "Linux Biolinum O"),
          axis.text.y  = element_text(size = 14),
          axis.text.x  = element_text(size = 14),
          axis.title   = element_text(size = 16),
          strip.text   = element_text(size = 20))

# Saving the plot
cairo_pdf("Figs/NonLin_Variances_cont.pdf", height = 5, width = 8)
plot(p_var)
dev.off()

# Statistical test for bias
tbl_var_plot |>
    summarise(P_val = wilcox.test(Bias)[["p.value"]],
              .by = c(Shape, Variance)) |>
    pivot_wider(values_from = "P_val", names_from = "Variance") |>
    mutate(across(-Shape, \(col) { num(col, digits = 5) })) |>
    print(width = Inf)
tbl_var_plot |>
    summarise(Mean_Bias = mean(Bias),
              Mean_Rel_Bias = mean(Rel_Bias),
              .by = c(Shape, Variance)) |>
    print(n = Inf)

# Correlation V_tot and V_P
tbl_vars |>
    summarise(Corr = cor(V_tot, V_phen),
              .by = c(Shape))

## Examples of reaction norms
## Sigmoid
tbl_annot_plas <-
    tbl_true_vars |>
    select(Shape, P2) |>
    pivot_longer(P2,
                 names_to = "Variance",
                 values_to = "Value") |>
    mutate(Variance = recode(Variance,
                             P2  = "P[RN]^2"),
           Value = round(Value, digits = 6) |>
                   signif(digits = 3) |>
                   map_chr(format, scientific = FALSE),
           Annot = str_c(Variance, " == ", Value),
           X     = -1.95,
           Y     = 0.9,
           Colour = set_colours(Variance),
           .by = Shape)

tbl_annot_pi <-
    tbl_true_vars |>
    select(Shape, starts_with("Pi")) |>
    pivot_longer(contains("Pi"),
                 names_to = "Pi",
                 values_to = "Value") |>
    mutate(X     = -1.95,
           Y     = c(0.75, 0.6),
           Pi = if_else(Pi == "Pi_Sl", "π[Sl]", "π[Cv]"),
           Value = round(Value, digits = 6) |>
                   signif(digits = 2) |>
                   map_chr(format, scientific = FALSE),
           Annot = str_c(Pi, " == ", Value),
           Colour = set_colours(Pi),
           .by = Shape)

tbl_annot_gen <-
    tbl_true_vars |>
    select(Shape, matches("^[Hh]"), starts_with("Gamma"), starts_with("Iota")) |>
    pivot_longer(-Shape,
                 names_to = "Parameter",
                 values_to = "Value") |>
    mutate(Parameter = case_when(Parameter == "H2_T"                            ~ "H[RN]^2",
                                 Parameter == "h2_T"                            ~ "h[RN]^2",
                                 Parameter == "h2_M"                            ~ "h^2",
                                 Parameter == "h2_I"                            ~ "h[I]^2",
                                 Shape == "Sigmoid"   & Parameter == "Gamma_1"  ~ "γ[L]",
                                 Shape == "Sigmoid"   & Parameter == "Gamma_2"  ~ "γ[r]",
                                 Shape == "Sigmoid"   & Parameter == "Gamma_12" ~ "γ[Lr]",
                                 Shape == "GausGompz" & Parameter == "Gamma_1"  ~ "γ[C]",
                                 Shape == "GausGompz" & Parameter == "Gamma_2"  ~ "γ[ε[0]]",
                                 Shape == "GausGompz" & Parameter == "Gamma_12" ~ "γ[Cε[0]]",
                                 Shape == "Sigmoid"   & Parameter == "Iota_1"   ~ "ι[L]",
                                 Shape == "Sigmoid"   & Parameter == "Iota_2"   ~ "ι[r]",
                                 Shape == "Sigmoid"   & Parameter == "Iota_12"  ~ "ι[Lr]",
                                 Shape == "GausGompz" & Parameter == "Iota_1"   ~ "ι[C]",
                                 Shape == "GausGompz" & Parameter == "Iota_2"   ~ "ι[ε[0]]",
                                 Shape == "GausGompz" & Parameter == "Iota_12"  ~ "ι[Cε[0]]") |>
                       factor(levels = c("H[RN]^2",
                                         "h[RN]^2",
                                         "γ[L]",
                                         "γ[r]",
                                         "γ[Lr]",
                                         "γ[C]",
                                         "γ[ε[0]]",
                                         "γ[Cε[0]]",
                                         "h^2",
                                         "h[I]^2",
                                         "ι[L]",
                                         "ι[r]",
                                         "ι[Lr]",
                                         "ι[C]",
                                         "ι[ε[0]]",
                                         "ι[Cε[0]]")),
           Value = round(Value, digits = 6) |>
                   signif(digits = 3) |>
                   map_chr(format, scientific = FALSE),
           Annot = str_c(Parameter, " == ", Value),
           Colour = set_colours(Parameter)) |>
    arrange(Shape, Parameter) |>
    mutate(X     = -2,
           Y     = seq(0.18, 0.03, length.out = n()),
           .by = Shape)

# Average phenotype for each environment
avg_phen_sig <-
    map_dbl(env_seq,
            compute_Eg_e,
            shape   = sigmoid_cuba,
            theta    = fe_sig,
            G_theta  = tbl_vars[["G"]][[1]],
            fixed   = NA,
            width = 9.5)

# Now generating the plot
p_sig <-
    ggplot() +
    geom_line(data = tbl_sim[["Data"]][[1]],
              aes(x = Env, y = Phen, group = Individual),
              colour = "grey",
              alpha = 0.5,
              linewidth = 1) +
    geom_function(fun = func_sig_1,
                  linewidth = 1.2,
                  colour = "black") +
    geom_line(data = NULL,
                aes(x = env_seq, y = avg_phen_sig),
                linewidth = 1.2,
                colour = set_colours("V[Plas]")) +
    annotate("rect", xmin = -3, xmax = -0.6, ymin = 1.15, ymax = 1.8,
         fill = "white", colour = "#FFCC00", linewidth = 2) +
    annotate("text", x = -2.25, y = 1.7, label = "italic(bar(L)) == 1",
             size = 6,
             family = "Linux Libertine O",
             parse = TRUE) +
    annotate("text", x = -1.3, y = 1.7, label = "italic(bar(r)) == 3",
             size = 6,
             family = "Linux Libertine O",
             parse = TRUE) +
    annotate("text", x = -1.8, y = 1.4, label = "G[paste(L,',',r)] == bgroup('(', atop('0.1   0.02', '0.02 0.05'),')')",
             size = 6,
             family = "Linux Libertine O",
             parse = TRUE) +
    annotate("rect", xmin = -3, xmax = -0.9, ymin = 0.5, ymax = 1.05,
             fill = "white", colour = "#FFCC00", linewidth = 2) +
    geom_text(data = tbl_annot_plas |> filter(Shape == "Sigmoid"),
              mapping = aes(x = X, y = Y, label = Annot, colour = I(Colour)),
              parse = TRUE,
              family = "Linux Biolinum O",
              size  = 6) +
    geom_text(data = tbl_annot_pi |> filter(Shape == "Sigmoid"),
              mapping = aes(x = X, y = Y, label = Annot, colour = I(Colour)),
              parse = TRUE,
              family = "Linux Biolinum O",
              size  = 6) +
    xlim(c(-3, 3)) +
    ylab("Phenotype") + xlab("Environment") +
    theme(legend.position = "none")

# Also generating the equation
p_eq_sig <-
    ggplot() +
    geom_text(aes(x = 0, y = 0.5, label = "Sigmoid"),
              family = "Linux Biolinum O",
              fontface = "bold",
              size = 10) +
    geom_text(aes(x = 0, y = 0, label = "italic(f(ɛ) == frac(L, 1 + e^{-r * ɛ}))"),
              size = 10,
              family = "Linux Libertine O",
              parse = TRUE) +
    ylim(c(-0.2, 0.6)) +
    theme_void() +
    theme(plot.margin = margin(0,0,0,0, "pt"))

## Performance curve

# Average phenotype in each environment
avg_phen_gg <-
    map_dbl(env_seq,
            compute_Eg_e,
            shape   = gausgompz_cuba,
            theta    = fe_gg,
            G_theta  = tbl_vars[["G"]][[n_sim + 1]],
            fixed   = fixed_gg,
            width   = 9.5)

# Now generating the plot
p_gg <-
    ggplot() +
    geom_line(data = tbl_sim[["Data"]][[n_sim + 1]],
              aes(x = Env, y = Phen, group = Individual),
              colour = "grey",
              alpha = 0.5,
              linewidth = 1) +
    geom_function(fun = func_gg_1,
                  linewidth = 1.2,
                  colour = "black") +
    geom_line(data = NULL,
                aes(x = env_seq, y = avg_phen_gg),
                linewidth = 1.2,
                colour = set_colours("V[Plas]")) +
    annotate("rect", xmin = -3, xmax = -0.6, ymin = 1.15, ymax = 1.8,
             fill = "white", colour = "#FFCC00", linewidth = 2) +
    annotate("text", x = -1.8, y = 1.7, label = "paste(italic(bar(C)) == 1, '    ', italic(bar(ε[0])) == 0.9)",
             size = 6,
             family = "Linux Libertine O",
             parse = TRUE) +
    annotate("text", x = -1.8, y = 1.55, label = "paste(italic(bar(ρ)) == 8, '    ', italic(bar(σ)) == 0.4)",
             size = 6,
             family = "Linux Libertine O",
             parse = TRUE) +
    annotate("text", x = -1.8, y = 1.33, label = "G[paste(C,',',ε[0])] == bgroup('(', atop('0.1   0.01', '0.01 0.05'),')')",
             size = 5.5,
             family = "Linux Libertine O",
             parse = TRUE) +
    annotate("rect", xmin = -3, xmax = -0.9, ymin = 0.5, ymax = 1.05,
             fill = "white", colour = "#FFCC00", linewidth = 2) +
    geom_text(data = tbl_annot_plas |> filter(Shape == "GausGompz"),
              mapping = aes(x = X, y = Y, label = Annot, colour = I(Colour)),
              parse = TRUE,
              family = "Linux Biolinum O",
              size  = 6) +
    geom_text(data = tbl_annot_pi |> filter(Shape == "GausGompz"),
              mapping = aes(x = X, y = Y, label = Annot, colour = I(Colour)),
              parse = TRUE,
              family = "Linux Biolinum O",
              size  = 6) +
    ylim(c(0, 1.8)) +
    xlim(c(-3, 3)) +
    ylab("Phenotype") + xlab("Environment") +
    theme(legend.position = "none")

# Also generating the equation
p_eq_gg <-
    ggplot() +
    geom_text(aes(x = 0, y = 0.5, label = "Thermal Perf. Curve"),
              family = "Linux Biolinum O",
              fontface = "bold",
              size = 10) +
    geom_text(aes(x = 0, y = 0, label = "italic(f(ɛ) == C * e^{-e^{ρ * (ɛ - ɛ[0]) - 6} - σ * (ɛ - ɛ[0])^2})"),
              family = "Linux Libertine O",
              size = 10,
              parse = TRUE) +
    ylim(c(-0.2, 0.6)) +
    theme_void() +
    theme(plot.margin = margin(0,0,0,0, "pt"))

# Computing the theoretical genetic variances for more values
# tbl_true_vgen <-
#     tibble(Env  = env_seq,
#            V_g_e_sig = map_dbl(Env,
#                                compute_vg_e,
#                                shape       = sigmoid_cuba,
#                                theta        = avg_sig,
#                                G_theta      = mat_sigm,
#                                .progress   = TRUE),
#            V_g_e_gg = map_dbl(Env,
#                               compute_vg_e,
#                               shape       = gausgompz_cuba,
#                               theta        = avg_gg,
#                               G_theta      = mat_gausgompz,
#                               fixed       = fixed_gg,
#                               .progress   = TRUE),
#            V_A_e_sig = map_dfr(Env,
#                                compute_va_e,
#                                shape       = d_theta_sigmoid_cuba,
#                                theta        = avg_sig,
#                                G_theta      = mat_sigm,
#                                .progress   = TRUE)[["V_A"]],
#            V_A_e_gg = map_dfr(Env,
#                               compute_va_e,
#                               shape       = d_theta_gausgompz_cuba,
#                               theta        = avg_gg,
#                               G_theta      = mat_gausgompz,
#                               fixed       = fixed_gg,
#                               .progress   = TRUE)[["V_A"]]) |>
#     pivot_longer(starts_with("V"),
#                  names_pattern = "(V_[gA]_e)_(.*)",
#                  names_to = c("Variance", "Shape"),
#                  values_to = c("Value")) |>
#     mutate(Shape = recode(Shape, sig = "Sigmoid", gg = "GausGompz"),
#            Env_numfac = 5.5 + Env * (9/4),
#            Variance = factor(Variance, levels = c("V_g_e", "V_A_e")))


## Genetic variances along environments

tbl_plot_vars_e <-
    tbl_vars_e |>
    mutate(Variance = recode(Variance,
                             V_A_e = "V[A,ε]",
                             V_g_e = "V[G,ε]"),
           Colour = set_colours(Variance)) |>
    arrange(Variance)

# Now generating the plot for Sigmoid
p_sig_var_e <-
    ggplot(tbl_plot_vars_e |> filter(Shape == "Sigmoid")) +
    geom_ribbon(aes(x = Env, ymin = Low, ymax = Up, fill = I(Colour)),
                alpha = 0.3) +
    geom_line(aes(x = Env, y = Mean, colour = I(Colour)),
              linewidth = 1.2) +
    annotate("rect", xmin = -3, xmax = -1, ymin = 0.02, ymax = 0.19,
             fill = "white", colour = "#FFCC00", linewidth = 2) +
    geom_text(data = tbl_annot_gen |> filter(Shape == "Sigmoid"),
              mapping = aes(x = X, y = Y, label = Annot, colour = I(Colour)),
              parse = TRUE,
              family = "Linux Biolinum O",
              size  = 6) +
    xlim(c(-3, 3)) +
    ylim(c(0, 0.19)) +
    xlab("Environment") + ylab("(Additive) Genetic Variance")

# Now generating the plot for the Thermal Perf. Curve
p_gg_var_e <-
    ggplot(tbl_plot_vars_e |> filter(Shape == "GausGompz")) +
    geom_ribbon(aes(x = Env, ymin = Low, ymax = Up, fill = I(Colour)),
                alpha = 0.3) +
    geom_line(aes(x = Env, y = Mean, colour = I(Colour)),
              linewidth = 1.2) +
    annotate("rect", xmin = -3, xmax = -1, ymin = 0.02, ymax = 0.19,
             fill = "white", colour = "#FFCC00", linewidth = 2) +
    geom_text(data = tbl_annot_gen |> filter(Shape == "GausGompz"),
              mapping = aes(x = X, y = Y, label = Annot, colour = I(Colour)),
              parse = TRUE,
              family = "Linux Biolinum O",
              size  = 6) +
    xlim(c(-3, 3)) +
    ylim(c(0, 0.19)) +
    xlab("Environment") + ylab("(Additive) Genetic Variance")

## Gamma along environments
tbl_plot_gamma_e <-
    tbl_gamma_e |>
    mutate(Estimate = case_when(Shape == "Sigmoid"   & Estimate == "Gamma_1"  ~ "γ[L]",
                                Shape == "Sigmoid"   & Estimate == "Gamma_2"  ~ "γ[r]",
                                Shape == "Sigmoid"   & Estimate == "Gamma_12" ~ "γ[Lr]",
                                Shape == "GausGompz" & Estimate == "Gamma_1"  ~ "γ[C]",
                                Shape == "GausGompz" & Estimate == "Gamma_2"  ~ "γ[ε[0]]",
                                Shape == "GausGompz" & Estimate == "Gamma_12" ~ "γ[Cε[0]]") |>
                      factor(levels = c("γ[L]",
                                        "γ[r]",
                                        "γ[Lr]",
                                        "γ[C]",
                                        "γ[ε[0]]",
                                        "γ[Cε[0]]")),
           Colour = set_colours(Estimate))

# Now generating the plot for Sigmoid
p_sig_gamma_e <-
    ggplot(tbl_plot_gamma_e |> filter(Shape == "Sigmoid", between(Env, -2, 2))) +
    geom_ribbon(aes(x = Env, ymin = Low, ymax = Up, linetype = Estimate, fill = I(Colour)),
                alpha = 0.3) +
    geom_line(aes(x = Env, y = Mean, linetype = Estimate, colour = I(Colour)),
              linewidth = 1.2) +
    xlim(c(-3, 3)) +
    ylim(c(-0.76, 1.23)) +
    scale_linetype_manual(name = NULL,
                          values = c("solid", "twodash", "dotted"),
                          labels = scales::parse_format()) +
    xlab("Environment") + ylab("γ-decomposition") +
    theme(legend.position = c(0.85, 0.15))


# Now generating the plot for the Thermal Perf. Curve
p_gg_gamma_e <-
    ggplot(tbl_plot_gamma_e |> filter(Shape == "GausGompz", between(Env, -2, 2))) +
    geom_ribbon(aes(x = Env, ymin = Low, ymax = Up, linetype = Estimate, fill = I(Colour)),
                alpha = 0.3) +
    geom_line(aes(x = Env, y = Mean, linetype = Estimate, colour = I(Colour)),
              linewidth = 1.2) +
    xlim(c(-3, 3)) +
    ylim(c(-0.76, 1.23)) +
    scale_linetype_manual(name = NULL,
                          values = c("solid", "twodash", "dotted"),
                          labels = scales::parse_format()) +
    xlab("Environment") + ylab("γ-decomposition") +
    theme(legend.position = c(0.85, 0.15))


## Full graph with reaction norms and variance estimations
# Relative errors of the variance partitions for Sigmoid
p_var_sig <-
    ggplot(tbl_var_plot |> filter(Shape == "Sigmoid", !(Variance %in% c("V[Res]", "V[Tot]")))) +
    geom_violin(aes(x = Variance, y = Bias, fill = I(Colour)),
                alpha = 0.5,
                scale = "width") +
    geom_hline(yintercept = 0) +
    stat_summary(aes(x = Variance, y = Bias),
                 geom = "point",
                 colour = "#444444",
                 size = 2,
                 fun = "mean") +
#     facet_wrap(~ Est_Type, scales = "free") +
    scale_x_discrete(labels = scales::parse_format()) +
#     scale_fill_hue(guide  = "none") +
    labs(x = "Variance component", y = "Error") +
    theme(text         = element_text(family = "Linux Biolinum O"),
          axis.text.y  = element_text(size = 14),
          axis.text.x  = element_text(size = 14),
          axis.title   = element_text(size = 16),
          strip.text   = element_text(size = 20))

# Relative errors of the variance partitions for the Thermal Perf. Curve
p_var_gg <-
    ggplot(tbl_var_plot |> filter(Shape == "Thermal Perf. Curve", !(Variance %in% c("V[Res]", "V[Tot]")))) +
    geom_violin(aes(x = Variance, y = Bias, fill = I(Colour)),
                alpha = 0.5,
                scale = "width") +
    geom_hline(yintercept = 0) +
    stat_summary(aes(x = Variance, y = Bias),
                 geom = "point",
                 colour = "#444444",
                 size = 2,
                 fun = "mean") +
#     facet_wrap(~ Est_Type, scales = "free") +
    scale_x_discrete(labels = scales::parse_format()) +
    scale_fill_hue(guide  = "none") +
    labs(x = "Variance component", y = "Error") +
    theme(text         = element_text(family = "Linux Biolinum O"),
          axis.text.y  = element_text(size = 14),
          axis.text.x  = element_text(size = 14),
          axis.title   = element_text(size = 16),
          strip.text   = element_text(size = 20))

p_full <-
    (p_eq_sig / p_sig / p_sig_var_e / p_sig_gamma_e / p_var_sig +
    theme(plot.margin = unit(c(0,30,0,0), "pt"))) +
    plot_layout(heights = c(1, 2, 2, 2, 2))  |
    (p_eq_gg / p_gg / p_gg_var_e / p_gg_gamma_e /p_var_gg) +
    plot_layout(heights = c(1, 2, 2, 2, 2)) +
    theme(axis.text.x = element_text(vjust = 0.5))

# Now we can generate the full graph
cairo_pdf("Figs/NonLin_Full_cont.pdf", width = 12, height = 20)
plot(p_full)
grid.draw(linesGrob(x = unit(c(0.5, 0.5), "npc"), y = unit(c(0.02, 0.98), "npc")))
dev.off()
